requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
appnope @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_46iquy4y0r/croot/appnope_1750774774351/work
asttokens @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_11i8cnwuxs/croot/asttokens_1743630449718/work
certifi==2025.11.12
charset-normalizer==3.4.4
comm @ file:///opt/miniconda3/conda-bld/comm_1763119153747/work
contourpy==1.3.2
cycler==0.12.1
debugpy @ file:///opt/miniconda3/conda-bld/debugpy_1762421637624/work
decorator @ file:///opt/miniconda3/conda-bld/decorator_1757341235559/work
exceptiongroup @ file:///opt/miniconda3/conda-bld/exceptiongroup_1761560357548/work
executing @ file:///opt/miniconda3/conda-bld/executing_1757061234733/work
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
idna==3.11
ipykernel @ file:///opt/miniconda3/conda-bld/ipykernel_1762946072004/work
ipython @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_2evnd74hdv/croot/ipython_1734548057210/work
jedi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_38ctoinnl0/croot/jedi_1733987402850/work
Jinja2==3.1.6
joblib==1.5.2
jupyter_client @ file:///opt/miniconda3/conda-bld/jupyter_client_1762943080419/work
jupyter_core @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_044bmfoejx/croot/jupyter_core_1751991384131/work
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline @ file:///opt/miniconda3/conda-bld/matplotlib-inline_1762779171547/work
mpmath==1.3.0
nest-asyncio @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_310vb5e2a0/croot/nest-asyncio_1708532678212/work
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079158/work
pandas==2.3.3
parso @ file:///opt/miniconda3/conda-bld/parso_1762781705608/work
pexpect @ file:///opt/miniconda3/conda-bld/pexpect_1762535936197/work
pillow==12.0.0
platformdirs @ file:///opt/miniconda3/conda-bld/platformdirs_1762356633662/work
polars==1.35.2
polars-runtime-32==1.35.2
prompt_toolkit @ file:///opt/miniconda3/conda-bld/prompt-toolkit_1761744860944/work
psutil @ file:///opt/miniconda3/conda-bld/psutil_1761896322276/work
ptyprocess @ file:///opt/miniconda3/conda-bld/ptyprocess_1762424170819/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=3470be7f810474c8a2ecfcd6e02acc6aea8483ab595417fa4e336362a349933e
pure_eval @ file:///opt/miniconda3/conda-bld/pure_eval_1757067086281/work
pydantic==2.12.4
pydantic_core==2.41.5
Pygments @ file:///opt/miniconda3/conda-bld/pygments_1762431566026/work
pyparsing==3.2.5
python-dateutil @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_66ud1l42_h/croot/python-dateutil_1716495741162/work
pytz==2025.2
PyYAML==6.0.3
pyzmq @ file:///opt/miniconda3/conda-bld/pyzmq_1762375617494/work
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.3
six @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_03myqm7p6o/croot/six_1744271511946/work
stack_data @ file:///opt/miniconda3/conda-bld/stack_data_1757067065720/work
stringzilla==4.2.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tornado @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_63s3osc0_1/croot/tornado_1748956943583/work
tqdm==4.67.1
traitlets @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_500m2_1wyk/croot/traitlets_1718227071952/work
typing-inspection==0.4.2
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_dffmcwqo6o/croot/typing_extensions_1756281471323/work
tzdata==2025.2
ultralytics==8.3.232
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_8el7txr34i/croot/wcwidth_1750352901095/work
